requirements.txt
pydantic>=2.7
pytz
python-dateutil
requests
pandas
numpy
ta
